move all internal usage to public api
