+++ nominal/cli/attachment.py
@@ -4,8 +4,8 @@ from pathlib import Path
 
 import click
 
-from nominal._utils import FileType
 from nominal.cli.util.global_decorators import client_options, global_options
+from nominal.core.filetype import FileType
 from nominal.core.client import NominalClient
